implement pre-issue access token action versioning and token exchange grant support
